+++ src/core/fields/mod.rs
@@ -1,4 +1,4 @@
 pub mod avx512_m31;
 pub mod cm31;
-pub mod ecm31;
 pub mod m31;
+pub mod qm31;
